+++ src/main/kotlin/controller/GameController.kt
@@ -2,10 +2,10 @@ package controller
 
 import ExceptionHandler
 import model.Answer
+import model.BlackjackGame
 import model.Hand
 import model.card.Deck
 import model.human.Dealer
-import model.human.Human
 import model.human.HumanName
 import model.human.Player
 import model.human.Players
@@ -13,6 +13,8 @@ import view.InputView
 import view.OutputView
 
 class GameController(private val deck: Deck) {
+    val blackjackGame = BlackjackGame(deck)
+
     fun start() {
         val dealer = Dealer(Hand())
         val players = readPlayers()
@@ -50,58 +52,25 @@ class GameController(private val deck: Deck) {
         dealer: Dealer,
         players: Players,
     ) {
-        initDealer(dealer)
-        initPlayers(players)
-        OutputView.showGameInit(dealer, players)
-    }
-
-    private fun initDealer(dealer: Dealer) {
-        initPlayer(dealer)
-    }
-
-    private fun initPlayers(players: Players) {
-        players.players.forEach { player ->
-            initPlayer(player)
-        }
-    }
-
-    private fun initPlayer(human: Human) {
-        repeat(2) {
-            human.hand.draw(deck.pop())
-        }
+        blackjackGame.initGame(dealer = dealer, players = players, OutputView::showGameInit)
     }
 
     private fun playGame(
         dealer: Dealer,
         players: Players,
     ) {
-        players.players.forEach(::playOfOnePlayer)
-        dealer.play(deck)
-
-        repeat(dealer.getCountOfAdditionalDraw()) {
-            OutputView.drawCardForDealer()
-        }
+        blackjackGame.play(dealer = dealer, players = players, ::playOfOnePlayer)
+        printDealerCardDraw(dealer)
     }
 
     private fun playOfOnePlayer(player: Player) {
-        while (playByAnswer(readAnswer(player.humanInfo.humanName), player));
+        while (blackjackGame.playByAnswer(readAnswer(player.humanInfo.humanName), player, OutputView::showPlayerHand));
     }
 
-    private fun playByAnswer(
-        answer: Answer,
-        player: Player,
-    ): Boolean {
-        val isBusted =
-            when (answer) {
-                Answer.YES -> {
-                    player.hand.draw(deck.pop())
-                    player.isPossible()
-                }
-
-                Answer.NO -> false
-            }
-        OutputView.showPlayerHand(player)
-        return isBusted
+    private fun printDealerCardDraw(dealer: Dealer) {
+        repeat(dealer.getCountOfAdditionalDraw()) {
+            OutputView.drawCardForDealer()
+        }
     }
 
     private fun readAnswer(humanName: HumanName): Answer {
@@ -117,17 +86,8 @@ class GameController(private val deck: Deck) {
         OutputView.showDealerHandWithResult(dealer)
         OutputView.showPlayersHandWithResult(players)
 
-        totalJudge(dealer = dealer, players = players)
+        blackjackGame.judgeWinningResult(dealer = dealer, players = players)
         OutputView.showResultHeader()
         OutputView.showTotalResult(dealer = dealer, players = players)
-    }
-
-    private fun totalJudge(
-        dealer: Dealer,
-        players: Players,
-    ) {
-        players.players.forEach { player ->
-            dealer.judge(player)
-        }
     }
 }
